Add hint text to content blocks without user editable fields.
This is a UI improvement to explain why some blocks don't have editable fields, and give a hint as to where the fields can be edited. In the case of global blocks, their fields can be changed on the main options page. For the case studies list, users will need to edit the case study posts.