Add sass-loader update
I had tried to update this before and it caused things to break, so I didn't include it in the initial commit. I've since updated it successfully on another site though, so I've come back here to try again and it's worked this time. It might be that something was cached before, that was causing issues.